backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP Clients
aiohttp==3.9.1
httpx==0.26.0
requests==2.31.0

# Configuration & Validation (compatible versions)
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Redis & Cache
redis==5.0.1

# AI/ML
numpy<2
transformers==4.36.0
torch==2.1.2
sentencepiece==0.1.99

# Authentication & Security
passlib==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# OAuth
google-auth==2.26.2
google-auth-oauthlib==1.2.0

# Task Queue
celery==5.3.4
flower==2.0.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Utilities
python-dateutil==2.8.2

# PDF Processing (optional, lightweight)
PyPDF2==3.0.1

# Text Processing
beautifulsoup4==4.12.2
lxml==5.1.0
